% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_addr.R
\name{prep_addr}
\alias{prep_addr}
\title{Remove potentially illegal characters from a vector or column of addresses.}
\usage{
prep_addr(addr_vector)
}
\arguments{
\item{addr_vector}{vector or data frame column of input addresses.  Required.}
}
\value{
A vector or column of addresses stripped of potentially illegal characters leaving only letters, numbers, hyphens, and spaces remaining.
}
\description{
The \code{prep_addr} function removes potentially illegal characters from a vector or column of addresses leaving only letters, numbers, hyphens, and spaces remaining.
}
\examples{
# create a data frame of addresses
FAC_NAME <- c("DOHMH","DOHMH","DOHMH","LEHMAN COLLEGE","BAM","DOHMH")
ADDR1 <- c("80 CENTER' ST","125 WORTH_","42-09 28 ST.",
    "250 BEDFORD P@RK BLVD W","#30 LAFAYETTE AVE","125~")
ADDR2 <- c("","UNIT #329","1$ FLR","212","ROOM 3","WORTH STREET")
BORO_CODE <- rep(1,length(ADDR1))
u_id <- 1:length(ADDR1)
df = data.frame(u_id, FAC_NAME, ADDR1, ADDR2, BORO_CODE)

#remove illegal characters from address column
df$ADDR1.prep <- prep_addr(df$ADDR1)
df$ADDR2.prep <- prep_addr(df$ADDR2)

#preview records
head(df)
}
