% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splchk_addr.R
\name{splchk_addr}
\alias{splchk_addr}
\title{Spell check a data frame of NYC addresses.}
\usage{
splchk_addr(in_df, new_addr_col_name, addr_col_name,
    third_col_name, third_col_type)
}
\arguments{
\item{in_df}{a data frame containing NYC addresses.  Required.}

\item{new_addr_col_name}{new_addr_col_name the name of output addresses column as string.  Required.}

\item{addr_col_name}{the name of the input addresses column as string.  Required.}

\item{third_col_name}{the name of either the borough code or zip code column as string.  Required.}

\item{third_col_type}{either \code{"boro_code"} or \code{"zip_code"} as string.  Required.}
}
\value{
A data frame containing the input data frame plus the spell checked address column.
}
\description{
The \code{splchk_addr} function performs a spell check on a data frame of NYC addresses with a street name dictionary built from NYC Department of City Planning's (DCP) PAD (Property Address Directory) and SND (Street Name Dictionary) files.
}
\examples{
# create a data frame of addresses
ADDR <- c("1212 AMESTERDAM AVEN","253 BROADWY",
    "250 BREDFORD PORK BLVD W","30 LAFAYET AVE")
CITY <- c("NEW YORK","NEW YORK","BRONX","BROOKLYN")
STATE <- rep("NY",length(ADDR))
ZIP_CODE <- c("10027","10007","10468","11217")
u_id <- 1:length(ADDR)
df = data.frame(u_id, ADDR, CITY, STATE, ZIP_CODE)

#get version of DCP PAD used to build package data
rNYCclean::pad_version

#spell check address column using zip code
df1 <- splchk_addr(df,"ADDR.splchk","ADDR","ZIP_CODE","zip_code")

#preview records
head(df1)
}
